+++ src/Plugin.cs
@@ -30,7 +30,7 @@ public class Plugin : BaseUnityPlugin
 {
   private const string ModGUID = "nbusseneau.LotusEcarlateChanges";
   private const string ModName = "LotusEcarlateChanges";
-  private const string ModVersion = "0.2.0";
+  private const string ModVersion = "0.3.0";
 
   public static new ManualLogSource Logger;
   public static Harmony Harmony;