+++ CHANGELOG.md
@@ -7,21 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Adjust armor recipes to use less leather from Monstrum creatures considering spawn rates.
+
+### Fixed
+
 - Fix `Chieftain deer head` recipe to properly require deer trophies.
 
 ## [0.2.0] - 2024-05-01
+
+### Added
 
 - Double the range of fireplace and torch-like prefabs to allow counterplay to the modpack's extremely dark nights.
 
 ## [0.1.2] - 2024-04-30
 
+### Changed
+
 - Remove `Assembly.LoadFile` calls in order not to trigger Thunderstore submission rejection filter.
 
 ## [0.1.1] - 2024-04-30
 
+### Fixed
+
 - Fix Thunderstore manifest missing dependencies.
 
 ## [0.1.0] - 2024-04-30
+
+### Added
 
 - Initial release.